Place a marker in the scene on map ctrl-click (using view._scene._nameSpace.doc.ctx.pickValue())
The problem is that after changing the viewpoint the scene must be
rendered before calling pickValue() and that I do not know how to use
an offscreen scene with x3dom.

The workaround I found is to change the viewpoint on ctrl-mousemove over the
map, so that when the map is clicked, the scene needed for pickValue() is
already rendererd and pickvalue cam return the proper coordinates.

TODO:
- change the viewpoint on keydown ctrl over the map (dont wait mousemove)
- implement code related to issue #4
- restore the original viewpoint when ctrl is released or
  on mouseleave map
- ...